Move AbstractNode in .impl package
