+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/JavadocElement.java
@@ -4,7 +4,7 @@
 
 package net.sourceforge.pmd.lang.java.ast;
 
-import net.sourceforge.pmd.lang.ast.AbstractNode;
+import net.sourceforge.pmd.lang.ast.impl.AbstractNode;
 import net.sourceforge.pmd.lang.java.javadoc.JavadocTag;
 
 public class JavadocElement extends AbstractNode {
